Fix automation condition labels in dg list command
The `dg list` command was displaying `None` for automation conditions without explicit labels (e.g., conditions created via `&` or `|` operators).

Changes:
- Moved `get_expanded_label` function from dagster-graphql to core dagster
  (serialized_objects.py) to make it accessible to both graphql and components
- Updated `dg list` to use `get_expanded_label()` instead of `get_label()`,
  which recursively builds meaningful labels from condition tree structure
- Updated dagster-graphql imports to use the relocated function

Result: `dg list` now shows readable labels like "AND (cond1) AND (cond2)" instead of `None` for composite automation conditions.
